Rename app arReticle -b com.myviroapp.arreticle
