+++ ARReticle/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.myviroapp",
+    package = "com.myviroapp.arreticle",
 )
 
 android_resource(
     name = "res",
-    package = "com.myviroapp",
+    package = "com.myviroapp.arreticle",
     res = "src/main/res",
 )
