+++ ARReticle/android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:tools="http://schemas.android.com/tools"
-  package="com.myviroapp">
+  package="com.myviroapp.arreticle">
     <queries>
         <package android:name="com.google.ar.core" />
     </queries>
