+++ ARReticle/android/app/src/main/java/com/myviroapp/arreticle/MainApplication.java
@@ -1,4 +1,4 @@
-package com.myviroapp;
+package com.myviroapp.arreticle;
 
 import android.app.Application;
 import android.content.Context;
@@ -64,7 +64,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.myviroapp.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.myviroapp.arreticle.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
