+++ ARReticle/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "myviroapp",
-  "displayName": "myviroapp"
+  "name": "arReticle",
+  "displayName": "arReticle"
 }
