+++ ARReticle/ios/arReticle.xcodeproj/xcshareddata/xcschemes/arReticle.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "myviroapp.app"
-               BlueprintName = "myviroapp"
-               ReferencedContainer = "container:myviroapp.xcodeproj">
+               BuildableName = "arReticle.app"
+               BlueprintName = "arReticle"
+               ReferencedContainer = "container:arReticle.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "myviroappTests.xctest"
-               BlueprintName = "myviroappTests"
-               ReferencedContainer = "container:myviroapp.xcodeproj">
+               BuildableName = "arReticleTests.xctest"
+               BlueprintName = "arReticleTests"
+               ReferencedContainer = "container:arReticle.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "myviroapp.app"
-            BlueprintName = "myviroapp"
-            ReferencedContainer = "container:myviroapp.xcodeproj">
+            BuildableName = "arReticle.app"
+            BlueprintName = "arReticle"
+            ReferencedContainer = "container:arReticle.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "myviroapp.app"
-            BlueprintName = "myviroapp"
-            ReferencedContainer = "container:myviroapp.xcodeproj">
+            BuildableName = "arReticle.app"
+            BlueprintName = "arReticle"
+            ReferencedContainer = "container:arReticle.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
