+++ ARReticle/ios/arReticle/Images.xcassets/AppIcon.appiconset/Contents.json
