+++ ARReticle/ios/arReticleTests/Info.plist
