+++ ARReticle/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "myviroapp",
+  "name": "arReticle",
   "version": "0.0.1",
   "private": true,
   "scripts": {
